// @flywave/flywave-terrain-datasource/src/quantized-terrain/quantized-stratum-mesh/decoder/types.d.ts
export interface ExtensionHeader {
    extensionId: number;
    extensionLength: number;
}
/**
 * Metadata structure
 */
export interface Metadata {
    jsonLength: number;
    json: Record<string, unknown>;
}
/**
 * Stratum voxel data
 */
export interface StratumVoxelData {
    id: string;
    index: number;
    start: number;
    end: number;
    neighbors: [number, number, number];
    material: number;
}
/**
 * Material definition
 */
export interface Material {
    color: ColorRGBA;
    uvTransform: [number, number, number, number];
}
/**
 * RGBA color definition
 */
export interface ColorRGBA {
    r: number;
    g: number;
    b: number;
    a: number;
}
/**
 * Fault point coordinates
 */
export interface FaultPoint {
    x: number;
    y: number;
    z: number;
}
/**
 * Fault profile data
 */
export interface FaultProfileData {
    id: string;
    name: string;
    type: string;
    strike: number;
    dip: number;
    throw: number;
    points: FaultPoint[];
}
/**
 * Trajectory point for boreholes
 */
export interface TrajectoryPoint {
    depth: number;
    x: number;
    y: number;
    z: number;
    azimuth: number;
    inclination: number;
}
/**
 * Borehole stratum layer
 */
export interface BoreholeStratum {
    id: string;
    lithology: string;
    top: number;
    base: number;
}
/**
 * Borehole data
 */
export interface BoreholeData {
    id: string;
    location: [number, number, number];
    depth: number;
    azimuth: number;
    inclination: number;
    trajectory: TrajectoryPoint[];
    stratums: BoreholeStratum[];
}
// ...
export interface CollapsePillarData {
    id: string;
    name: string;
    topCenter: [number, number, number];
    baseCenter: [number, number, number];
    topRadius: number;
    baseRadius: number;
    height: number;
    stratumId: string;
    lithology: string;
}
/**
 * Section line data
 */
export interface SectionLineData {
    id: string;
    name: string;
    lineString: Array<[number, number, number]>;
}